Tidy JavaScript and add comments
In this commit, I mostly modified the Javascript to make it a bit more
terse, and added a few things. Warning messages now appear above the
element they are related to. Creation of the round 1 bracket is now
simpler since I removed a useless array.

I also did a few style things and rewordings, but the average person
would probably not notice. All in all, the whole thing is a bit prettier
and better worded. What a great day.

~ Love, Ted M.