Fix a minor bug and a typo
There were two bugs that still passed through the last commit. One was an error of using `append` rather than `appendChild` in the script. The second was the README.adoc saying that brackets up to 16 players could be created. We can now create up to 32 player brackets properly. So both of those were fixed. This is the last commit today, so good night, or good morning, however you look at it.

~ Love, Ted M.